// frontend/src/api-client/models/TenantBranchDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TenantBranchDto = {
    id?: string;
    tenantId?: string;
    code?: string;
    nameVi?: string;
    addressLine?: string;
    city?: string;
    district?: string;
    ward?: string;
    phone?: string;
    isActive?: boolean;
    createdAt?: string;
    updatedAt?: string;
};