// frontend/src/api-client/models/CreateBranchRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateBranchRequest = {
    tenantId: string;
    code: string;
    nameVi: string;
    addressLine?: string;
    city?: string;
    district?: string;
    ward?: string;
    phone?: string;
};